scripts/service:  introduce cronjobs as an optional feature
this is more of an automation / advanced feature
if this is planned to be on the webui, it likely
has to be added behind developer mode.

I really don't want to explain to people on how
cronjobs work and how to use it.

to enable

sh /data/adb/modules/bindhosts/enable_cron.sh

I also don't plan to give support on this feature.
none of that "sir, how do I run the cronjob on my birthday"
basically, its there, but not supported.