scripts: handle mode 1 restoration on legacy susfs
since we can't do kstat spoof anymore due to
https://gitlab.com/simonpunk/susfs4ksu/-/commit/3bdce5608395dd1c7dffcdee89d416e3b88375fc

while mode 8 is good and all for legacy susfs, it is still preferrable
to use mode 1. this handles it conditionally. basically if 153 or newer,
no more kstat spoofing is needed. for 152 and older, kstat spoofing is
still needed on some setups weirdly enough.