scripts/post-fs-data: probe mode 2 for APatch litemode and MKSU nomount
With APatch and Magic-KSU now supporting a lite/nomount mode, this is the only
way this module can maintain systemless hosts functionality when that mode is enabled.

ofcourse, if user has other stuff e.g. susfs, hosts_file_redirect, ZN-hostsredirect,
that will be used as shell scripts do in-order-exec.

References:
	https://github.com/5ec1cff/KernelSU/commit/76bfccd11f4c8953b35e1342a2461f45b7d21c22 /data/adb/ksu/.nomount
	https://github.com/bmax121/APatch/commit/7760519b89917a4f39f6b64c95fc6fa311da264d /data/adb/.litemode_enable

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>